+++ client/src/input/Input.js
@@ -453,6 +453,17 @@ export default Garnish.Base.extend({
     return this._blocks.filter(block => block.$container.closest($selectedBlocks).length > 0)
   },
 
+  getCopiedBlocks () {
+    const copyData = window.localStorage.getItem(`neo:copy:${this._name}`)
+
+    if (!copyData) {
+      return []
+    }
+
+    const { blocks } = JSON.parse(copyData)
+    return blocks
+  },
+
   _setMatrixClassErrors () {
     // TODO: will need probably need to find a method within php instead of JS
     // temp solution for now.
@@ -639,6 +650,11 @@ export default Garnish.Base.extend({
   },
 
   _getNextBlockIndex (index) {
+    // If undefined, then there's no previous block and the 'next' block will be the first block
+    if (typeof index === 'undefined') {
+      return 0
+    }
+
     if (index instanceof Block) {
       index = this._blocks.indexOf(index)
     }
@@ -655,7 +671,11 @@ export default Garnish.Base.extend({
     const animate = !Garnish.prefersReducedMotion()
     const $spinner = $(`<div class="ni_spinner">${animate ? '<div class="spinner"></div>' : 'Loading block'}</div>`)
 
-    block.$container.after($spinner)
+    if (typeof block !== 'undefined') {
+      block.$container.after($spinner)
+    } else {
+      this.$blocksContainer.prepend($spinner)
+    }
 
     let spinnerComplete = false
     let spinnerCallback = function () {}
@@ -864,7 +884,7 @@ export default Garnish.Base.extend({
       }
     }
 
-    window.localStorage.setItem('neo:copy', JSON.stringify(data))
+    window.localStorage.setItem(`neo:copy:${this._name}`, JSON.stringify(data))
 
     this._updateButtons()
 
@@ -874,12 +894,10 @@ export default Garnish.Base.extend({
 
   '@pasteBlock' (e) {
     const block = e.block
-    const baseLevel = block.getLevel() - 1
-    const copyData = window.localStorage.getItem('neo:copy')
+    const baseLevel = (block?.getLevel() ?? 1) - 1
+    const blocks = this.getCopiedBlocks()
 
-    if (copyData) {
-      const { blocks } = JSON.parse(copyData)
-
+    if (blocks.length > 0) {
       for (const pasteBlock of blocks) {
         pasteBlock.level += baseLevel
       }